project/ScaledYOLOv4/requirements.txt
onnx ==1.6.0
pytorch >= 1.6.0
onnx-simplifier
